# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Auto Router API
  version: 0.10.0
  summary: Single-user LLM gateway with named inference keys, provider configuration, fallback groups and usage metrics.
  description: |
    Auto Router is a single-user model gateway with provider-level upstream credentials, named client
    inference keys, and ordered fallback groups (simple, medium and complex). The first owner password
    is set once through the local setup endpoint; later access uses a password-backed browser session.

    Management APIs use the `auto_router_admin_session` cookie. Inference calls use a named inference
    key as `Authorization: Bearer <key>`. Provider secrets are write-only. Groups reference enabled
    provider/model pairs, and each group's array order is the fallback order.

    OpenAI-compatible ingress is available at `/v1/chat/completions` and `/v1/responses`; native
    Anthropic ingress is available at `/v1/messages`, and native Gemini ingress at
    `/v1beta/models/{model}:generateContent`. Native Anthropic and Gemini ingress requires the
    matching upstream protocol. OpenAI Chat text-only requests can be translated to Anthropic or
    Gemini for a limited supported subset; unsupported fields are rejected before upstream delivery.

    `GET /admin/v1/logs/summary` reports final client request success rate, token observations and
    60-second completed upstream output-token TPS. `GET /admin/v1/logs/stats` can aggregate actual
...
  license:
    name: MIT
...
servers:
...
security: []
paths:
...
  /admin/v1/setup/status:
    get:
      tags: [Setup]
      summary: Check whether the owner password has been set
      operationId: getSetupStatus
      security: []
      responses:
        '200':
          description: Setup status.
          content:
            application/json:
              schema:
                type: object
                required: [password_set]
                properties:
                  password_set: {type: boolean}
        '503':
          $ref: '#/components/responses/AdminError'
  /admin/v1/setup/password:
    post:
      tags: [Setup]
      summary: Set the initial owner password
      operationId: setInitialPassword
      description: Available only before a password exists and only from loopback with a same-origin request. Password length is 12–1024 bytes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: {type: string, minLength: 12, maxLength: 1024, writeOnly: true}
      responses:
        '201':
          description: Initial password set.
          content:
            application/json:
              schema:
                type: object
                required: [password_set]
                properties:
                  password_set: {type: boolean, const: true}
        '400': {$ref: '#/components/responses/AdminError'}
        '403': {$ref: '#/components/responses/AdminError'}
        '409': {$ref: '#/components/responses/AdminError'}
        '429': {$ref: '#/components/responses/AdminError'}
        '503': {$ref: '#/components/responses/AdminError'}
  /admin/v1/password:
    patch:
      tags: [Admin]
      summary: Change the owner password
      operationId: changeOwnerPassword
      security:
        - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, password]
              properties:
                current_password: {type: string, writeOnly: true}
                password: {type: string, minLength: 12, maxLength: 1024, writeOnly: true}
      responses:
        '200':
          description: Password changed; existing browser sessions are revoked.
          content:
            application/json:
              schema:
                type: object
                required: [changed]
                properties:
                  changed: {type: boolean, const: true}
        '400': {$ref: '#/components/responses/AdminError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/InsufficientScope'}
  /admin/v1/dashboard:
    get:
      tags: [Admin]
      summary: Read dashboard usage metrics
      operationId: getAdminDashboard
      security:
        - adminSession: []
      parameters:
        - name: window
          in: query
          required: false
          description: Reporting period, defaulting to 24h.
          schema: {type: string, enum: [1h, 24h, 7d, 30d], default: 24h}
      responses:
        '200':
          description: Final client success rate, model and group upstream attempt usage, and output TPS.
          content:
            application/json:
              schema:
                type: object
                required: [window, success_rate, models, groups, output_tps_60s]
                properties:
                  window:
                    type: object
                    required: [key, from, to]
                    properties:
                      key: {type: string, enum: [1h, 24h, 7d, 30d]}
                      from: {type: string, format: date-time}
                      to: {type: string, format: date-time}
                  success_rate: {$ref: '#/components/schemas/DashboardMetric'}
                  models:
                    type: array
                    items: {$ref: '#/components/schemas/DashboardCountRow'}
                  groups:
                    type: array
                    items: {$ref: '#/components/schemas/DashboardCountRow'}
                  output_tps_60s: {type: [number, 'null'], format: double}
        '400': {$ref: '#/components/responses/AdminError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/InsufficientScope'}
        '500': {$ref: '#/components/responses/AdminError'}
  /admin/v1/groups:
    get:
      tags: [Admin]
      summary: Read the fixed routing model groups
      operationId: getModelGroups
      security:
        - adminSession: []
      responses:
        '200':
          description: Three ordered provider/model fallback lists.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelGroups'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/InsufficientScope'}
    put:
      tags: [Admin]
      summary: Replace all routing model groups
      operationId: putModelGroups
      description: All three groups are required. Array order is fallback order; each group accepts up to eight distinct enabled provider/model pairs.
      security:
        - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModelGroups'}
      responses:
        '200':
          description: The saved ordered groups.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelGroups'}
        '400': {$ref: '#/components/responses/AdminError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/InsufficientScope'}
...
  /admin/v1/models:
    get:
      tags: [Admin]
      summary: List logical models
      operationId: listAdminModels
      description: |
        Returns logical models from the database (the authority the published snapshot is built
        from) in contract order: priority ascending, then ID ascending. `next_cursor` is an opaque
        position key rather than an offset, so a concurrent write cannot make a row appear twice or
        disappear.
      security:
        - adminSession: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - name: enabled
          in: query
          required: false
          description: Filter by the row's enabled flag. Anything other than a boolean is `400 invalid_filter`.
          schema: {type: boolean}
        - name: source
          in: query
          required: false
          description: Filter by where the row came from. `local` is an administrator-owned row; `modelsdev` came from a models.dev synchronization.
          schema:
            type: string
            enum: [local, modelsdev]
      responses:
        '200':
          description: One page of models.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPage'
        '400':
          $ref: '#/components/responses/AdminError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '500':
          $ref: '#/components/responses/AdminError'
    post:
      tags: [Admin]
      summary: Create a logical model
      operationId: createAdminModel
      description: |
        Inserts an administrator-owned local model. The `auto` identifier and the reserved
        `provider:` prefix are rejected, so no registry entry can shadow either request syntax.

        There is no corresponding `DELETE`: a model is disabled, never removed, so a stored log row
        never references a row that vanished.
      security:
        - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModelCreate'}
      responses:
        '201':
          description: The created model, as it will be reported from now on.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Model'}
        '400':
          $ref: '#/components/responses/AdminError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '409':
          description: A model with this ID already exists (`model_exists`).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminError'}
        '413':
          $ref: '#/components/responses/AdminError'
        '500':
          $ref: '#/components/responses/AdminError'
...
  /admin/v1/providers/{key}/discover:
    get:
      tags: [Admin]
      summary: Discover available upstream models
      operationId: discoverProviderModels
      security:
        - adminSession: []
      parameters:
        - $ref: '#/components/parameters/ProviderKey'
      responses:
        '200':
          description: Discovered model identifiers; discovery does not add them to the registry.
          content:
            application/json:
              schema:
                type: object
                required: [items, truncated]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, label]
                      properties:
                        id: {type: string}
                        label: {type: string}
                  truncated: {type: boolean}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/InsufficientScope'}
        '404': {$ref: '#/components/responses/NotFound'}
        '502': {$ref: '#/components/responses/AdminError'}
...
  /admin/v1/sync-state:
    get:
      tags: [Admin]
      summary: Report the last models.dev synchronization
      operationId: getAdminSyncState
      description: |
        Reports the stored outcome of the last models.dev synchronization. "Never synchronized" is a
        fact worth a `200` (`{"synchronized": false}`): the remedy is the explicit sync action, and a
        `404` would wrongly suggest the endpoint itself is missing.
      security:
        - adminSession: []
      responses:
        '200':
          description: The synchronization state.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SyncState'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '500':
          $ref: '#/components/responses/AdminError'
...
components:
  securitySchemes:
    inferenceKey:
      type: http
      scheme: bearer
      description: |
        A named inference key sent as `Authorization: Bearer <key>`.
    adminSession:
      type: apiKey
      in: cookie
      name: auto_router_admin_session
      description: |
        The browser session issued by `POST /admin/v1/session`. Accepted only under `/admin/`; it
        cannot authenticate inference requests.
...
  responses:
...
    NotFound:
      description: The path is not registered in this process (for example an opt-in debug endpoint that is disabled, or the management surface when it is not mounted).
...